Super bad use of git but I don't want to deal with the manual merge
so I'm fast-forwarding my master to match intellij-project-files branch
to get my changes into master, since the separate branch was an awful
idea. I've added the IntelliJ project files to the ignore list so they
shouldn't show up in master.

The version of WhatAndroid in bin is probably ready for a wider testing
release but I'm not sure, I'll have to stick it on my phone and fiddle
for a bit