Changing how versioning is treated and checking for updates
We now will increment versionCode each release and use versionName for
the release name, aka. 1.1, 1.2 etc. where versioning is Major.Minor

The UpdateChecker now checks on the Github Releases page and looks for a
higher version name tagged release than the current one. So all
release's tags should be their corresponding version name. Altough I
kind of goofed this up for the stable "release" and tagged it 1.01 but
it should be ok since it's still just the lowest version tag.

To download the update the user gets an alert dialog and when clicking
update the browser is opened to the releases page for that release so
they can see a change-log and the download button.