Implementing a rough version of the saved barcode searcher
Displays a list of the barcodes which can be clicked to load their
search terms, then  search torrents and search requests buttons become
available which will launch the corresponding search intent with the
search string. An option is also available to remove the entry from your
barcode file.

I'll be doing some more testing but using some dummy test barcodes
everything ran smoothly on api 10, 16 and 17 which I think is enough for
a wider testing preview release of this version (1.3).