Moving all work to the adapter
The adapter now sets a listener for each list view item (the list
entries) and these listeners will launch the term lookup task. Also
adding view recyclng and the viewholder pattern. We also use the
viewholder in the async task to display a loading spinner in the item
while we're loading the terms so we don't look unresponsive.

Still having one issue though in that the list items won't highlight
when you click them