Initial draft of fix for viewing non-music torrents
Crash was caused by attempting to use and set TextViews that didn't exist
in the view so I've chosen to create a custom layout for non-music torrents
that they'll use to setup their information.

I have tested this on 4.3 and will test on the other supported API levels
to make sure everything's patched up.