Have setup a basic announcements view + detail view for announcement
When logging in a list of announcements will be the first thing shown. Selecting an announcement adds a fragment displaying the full announcement body and slides to it. Blog posts will also be
handled in the same way by the same activity by swapping the AnnouncementsPagerAdapter for a BlogPostPagerAdapter, and selected via the nav-bar. The other options on the navbar
will go to their own activities.