Remove WhatAPI jar from repository
Since I've decided to keep the WhatAPI repo up to date on releases when
I make changes I plan to use in the app there's no need to keep the
binary in the repository and update it as well. Added instructions to
the readme about our dependencies and to the todo in libs.