Now properly handle various paranoia settings (sort of)
We need to do a lot of null checking since there's tons of various possible paranoia configurations which can cause us to receive nulls in the fields we're trying to display. Since the paranoia number itself isn't very informative (it's just the number of fields blocked) we can't base much off of it, or tell if the recent torrents/snatches specifically have been blocked. This is partly related to a bug in the API where user recents doesn't respect paranoia but that will likely be fixed soonish and then we'll need to add null checks for these fields. For now we hide them if the user's paranoia is High+ (ie 6+). We also need to do some checks and ignore the paranoia number if we're viewing our own profile, since we can always see our own stuff.