Adding basic Torrent Group view and shortcutting to test it
It's an identical themed layout to the Artist Activity, a big header image of the album art and a list of the all the torrents. Later the torrents will be grouped up by release to distinguish deluxe/special editions whatever. I need to double check that the ordering we get back for these is always correct though since it seems the API just throws us one huge list of torrents. If they aren't grouped by their release type and in chronological order we'll need to do the sorting/grouping on our end, or add a function to do it in WhatAPI