Adding TorrentSearch
Displays the same torrent group listing that the artist view does, but showing the search results of course. Also fixing a conflict between the manifest and gradle build settings.
The search activity will run whichever fragment search type we want to view, ie. Torrent, User, Request, Artist which will handle the loading and paging.

Todo is turn the TorrentSearchAdapter into an endless scrolling adapter by loading the next page (if there is one) as the user reaches the bottom of the current page.