The torrent search fragment now performs the initial load of the search
The adapter will only load the next pages (if there are any). This makes it possible to start loading a search prior to creating any of the views or doing any other setup if we're loading a search from an intent. Otherwise we just wait until the user starts a search with the button as usual. This is added so that the Artist search can behave like the site does, ie. jump to the artist if they exist or fall back to a torrent search if not. The adapter's click listener also takes into account the footer