Add Artist search functionality, behaves like the site
We try to load an artist by name, if we succeed we go to that artist, if we fail we run a torrent search with the name we tried to look up. TODO: pass the Artist through the intent as JSON instead of having the artist view reload the artist, since by doing the lookup in Artist.fromName we load the same type of artist object that the ArtistActivity loads.