Pass the loaded search result artist to the artist activity
When we do an artist search we try to load the Artist from name and if we succeed we've loaded the same artist type that the ArtistFragment displays, so instead of reloading it we make the loaded Artist available to the ArtistActivity and tell it to use the existing search artist. If there is no search artist (use search is false) it will default to loading the artist with some id.