Starting to work on user preferences and cookie saving
Also considering dropping support for API<8. We'll save the user's cookie (logged in with keep logged) so that we can jump into any of the activities and not worry about re-logging in. Will also provide a logout option to clear the saved cookies and such. Need to also add not-logged/no-cookie detection to the activities to handle it properly. ie. redirect to login activity or load the cookie.