Working on figuring out how to handle jumping into activities while not logged in
I think this works reasonably well but could probably be cleaned up some. A parent activity LoggedInActivity checks to make sure the user is logged in properly & has their index loaded and if they aren't we kick them to the login activity which will get them setup and return back to us. Once we're properly logged in we alert the child activity that it's ok to start making API calls and loading data, and it will pass on this message to whatever fragments/things that need API data it has.