Working on improving the Announcements view and fixing some of the performance issues
I think a lot of the performance problems come from displaying such large strings of text. Planning to add a get snippet method to the API that will get shorter snippets that we can show in the list view, but I'm not sure how much that'll really help performance. Maybe if we also stripped out the html characters, speaking of which I'm still unsure on how I want to handle them.