Adding error logging and reporting
The error logger catches uncaught exceptions, dumps an error report then passes the exception on to the original uncaught exception handler. The error reporter service is run on login and checks for any crash report files. If crash report files are found the user is given a notification to send the reports to the devs.