Some artists the cost of serializing is more/equal to the cost of re-downloading them
Artists with tons of releases (eg. Daft Punk, Kanye West, etc) have a ton of data to put in the bundle and when we save/reload it the app is unresponsive for a few seconds since it takes so long. Catch these cases and don't save them. The cutoff value may need some fine tuning to find a good spot.