Complete reworking of the TorrentGroup views
Couldn't think of a good way to show detailed torrent information
with the current layout so I've switched to a master/detail view
with the torrent group being the master and the torrent views being
the detail. Also added a comments listing fragment to the right of
the torrent group, but layout needs to be polished up.

Also added ability to click on a torrent group image and have it show
a pop-up of the image with no scaling applied so you can see the entire
album art. Will put downloading back in but will be done through the
torrent detail view instead, perhaps as an action bar item