Fiddling around with site emoticon to emoji support
I'm not sure about the speed of it but we do it in the background loader so we can probably get away with it. Also reducing the rate limit time out on the comments and adding support for basic Html tags with Html.fromHtml. I'm not sure what we should do with images in the comments, I'd like to load them (same with forums, etc) but as a temporary thing we hide them currently.