Change to having the art container be  a fixed size to prevent layout shuffling when showing the image
Since the spinner is smaller than the art displayed when the loading finished the bigger art would case the other items in the layout to shuffle around so it could fit. To prevent this we instead fix the size of the art container, which also results in the spinner being placed better as well (more centered).