Fix handling of abbreviated internal site links
For example the site allows links to be posted as just artist.php?id=x
but to pick these up with our intent filters we need the full site link
so expand it back out. I think these are the only links that will
not be starting with http so it should be fine.