Thinking about a new layout for the torrent/user/request searches
Instead of having the search boxes be in the list header and thus be scrolled away from as we browse we'd put them in a fixed view above the list that will collapse when we scroll past. Clicking this will then open up the search boxes so we can run a new search without needing to scroll all the way back up. This isn't implemented yet but I'd like to try it out after the forums are in and I start doing more polish work.