Can now actually vote on and view poll results
Changing to the result view after voting on a poll being viewed is still in progress. Instead of re-loading the entire page we'll simply update the poll on our side and show these new results. Also need to make it so the dialog changes to the results view instead of going away after clicking vote/show results.