Add basic support for subscriptions
You can now view unread (or all) subscribed forum threads,  grouped by the forum category they're under in the subscriptions activity. Clicking a thread will jump you to the last read post. We track the subscriptions and show the information the same way as notifications, eg. through a boolean preference that we update when loading the user's index and show the "new subscriptions" alert in the navbar based on this preference.

Todo: Custom layout for the subscription items so it looks nicer, along with support for unsubscribing from threads from the subscriptions view.