Tweaking NavDrawer some
I need a better method for finding and updating menu items so I've added some minor extensions to ArrayAdapter for the NavDrawer. This lets us do a fuzzy find and replace that's needed for updating the Notifications and Subscriptions menu items and now with this we can show the number of torrent notifications instead of just "new".