Add basic support for replying to threads
The fragment communicates with the thread fragment to request drafts be saved/discarded or that the current text be posted to the thread. The draft saving/restoring lets us compose or post while also being able to look back at the thread and add more user quotes or check the thread contents without losing our current composition.

Todo: quoting users and updating the forum thread shown after the reply is posted. We should also show a loading indicator when we launch the post reply task.

This same ReplyDialogFragment can be used for composing messages in the inbox as well when we start working on that section.