Don't save the user's password it's not secure
The SharedPreferences isn't as secure as I thought it is and even if we encrypt and store the password it really isn't that much better. The best approach is to only hang onto the cookie and user name. The user name we keep to tell if a different user is logging in so we can toss the old cookie. Then if the cookie expires we ask them to re-login.

I'm not really sure what to do about PyWA passwords, since PyWA doesn't support HTTPS we have to send it plain across the network, so it seems like it's a waste of time to bother.

The app will now also update the saved message count when logging in.