Fix bug when changing light preference in settings from the forum
If we navigate to the settings when viewing a forum page and change our light layout preference the forum listing we go back to won't update. This is resolved by checking the state in onResume instead, which will be called when coming back from another activity/fragment.

Also updated AS, seems to have tweaked project files some, and the autoformatter is tweaking some stuff.