Fix sending to Transdroid
Turns out that Transdroid has issues with content://downloads/my_downloads/#### uris and can only handle content://downloads/all_downloads/#### which we can't send. Solution is to get the local filename, parse the Uri for that and send that over so Transdroid can get it's act together.

Also remove sending of torrent id through constructor and instead parse it from the download uri which will be more robust since we only depend on the download data.