Semi-hacky fix for API < 23, seems to not pick a -night or -notnight
It seems that on some 5.1.1 devices the Appcompat lib or something isn't
choosing a theme to start with so we don't pick values-night or
values-notnight and thus are missing colors like background and such
when we try to draw the profile.

This hacky fix copies the light theme colors into values/colors.xml so
they'll be found without having to pick a theme. From there the user can
pick the light or dark theme as desired